requirements.txt
alembic==1.10.3
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
autopep8==2.0.2
bcrypt==4.0.1
billiard==3.6.4.0
bleach==6.0.0
blinker==1.6
cachelib==0.9.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-SSE==1.0.0
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mkdocs-material-extensions==1.1.1
passlib==1.7.4
prompt-toolkit==3.0.38
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
six==1.16.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.40.0
tomli==2.0.1
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
zope.event==4.6
zope.interface==6.0
